// src/app/precisa/proyecto/Programa/model/FiltroPrograma.ts
export class FiltroPrograma{

    constructor(){
    }

    IdProyecto:     number;
    CompaniaCodigo: string;
    Codigo:         string;
    Nombre:         string;
    IdSede:         number;
    Departamento:   string;
    Provincia:      string;
    Distrito:       string;
    FechaCreacion:       Date;
    FechaRegistro:       Date;
    Estado:         number;

}